consolidate bylaw models, simplify import parking data mgmt command
