+++ parking/tests/test_models.py
@@ -1,8 +1,7 @@
 from django.test import TestCase
 
 from whereToPark.models import (
-    NoParkingByLaw,
-    RestrictedParkingByLaw,
+    ByLaw,
     Highway,
     Intersection,
 )
@@ -30,13 +29,8 @@ class ByLawModelTest(TestCase):
             status="FNF",
         )
 
-
-class NoParkingByLawModelTest(ByLawModelTest):
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
         # setup objects used by all test methods
-        NoParkingByLaw.objects.create(
+        ByLaw.objects.create(
             source_id="1",
             schedule="15",
             schedule_name="Parking for Restricted Periods",
@@ -45,130 +39,63 @@ class NoParkingByLawModelTest(ByLawModelTest):
             boundary_end=cls.intersection_2,
             side="North",
             between="Dowling Avenue and Jameson Avenue",
-            prohibited_times_and_or_days="12 hours",
-        )
-
-    def test_source_id(self):
-        law = NoParkingByLaw.objects.get(id=1)
-        field_label = law._meta.get_field("source_id").verbose_name
-        self.assertEqual(field_label, "source id")
-
-    def test_schedule_label(self):
-        law = NoParkingByLaw.objects.get(id=1)
-        field_label = law._meta.get_field("schedule").verbose_name
-        self.assertEqual(field_label, "schedule")
-
-    def test_schedule_name_label(self):
-        law = NoParkingByLaw.objects.get(id=1)
-        field_label = law._meta.get_field("schedule_name").verbose_name
-        self.assertEqual(field_label, "schedule name")
-
-    def test_highway_label(self):
-        law = NoParkingByLaw.objects.get(id=1)
-        field_label = law._meta.get_field("highway").verbose_name
-        self.assertEqual(field_label, "highway")
-
-    def test_side_label(self):
-        law = NoParkingByLaw.objects.get(id=1)
-        field_label = law._meta.get_field("side").verbose_name
-        self.assertEqual(field_label, "side")
-
-    def test_between_label(self):
-        law = NoParkingByLaw.objects.get(id=1)
-        field_label = law._meta.get_field("between").verbose_name
-        self.assertEqual(field_label, "between")
-
-    def test_prohibited_times_and_or_days_label(self):
-        law = NoParkingByLaw.objects.get(id=1)
-        field_label = law._meta.get_field("prohibited_times_and_or_days").verbose_name
-        self.assertEqual(field_label, "prohibited times and or days")
-
-    def test_boundary_start_label(self):
-        law = NoParkingByLaw.objects.get(id=1)
-        field_label = law._meta.get_field("boundary_start").verbose_name
-        self.assertEqual(field_label, "boundary start")
-
-    def test_boundary_end_label(self):
-        law = NoParkingByLaw.objects.get(id=1)
-        field_label = law._meta.get_field("boundary_end").verbose_name
-        self.assertEqual(field_label, "boundary end")
-
-    def test_str_method(self):
-        law = NoParkingByLaw.objects.get(id=1)
-        self.assertEqual(law.__str__(), "queen street (North) - 1")
-
-
-class RestrictedParkingByLawModelTest(ByLawModelTest):
-    @classmethod
-    def setUpTestData(cls):
-        # setup objects used by all test methods
-        super().setUpTestData()
-        RestrictedParkingByLaw.objects.create(
-            source_id="1",
-            schedule="15",
-            schedule_name="Parking for Restricted Periods",
-            highway=cls.highway,
-            side="North",
-            between="Glenholme Avenue and Oakwood Avenue",
-            times_and_or_days="10:00 a.m. to 6:00 p.m., Mon. to Fri.",
+            times_and_or_days="12 hours",
             max_period_permitted="12 hours",
-            boundary_start=cls.intersection_1,
-            boundary_end=cls.intersection_2,
         )
 
     def test_source_id(self):
-        law = RestrictedParkingByLaw.objects.get(id=1)
+        law = ByLaw.objects.get(id=1)
         field_label = law._meta.get_field("source_id").verbose_name
         self.assertEqual(field_label, "source id")
 
     def test_schedule_label(self):
-        law = RestrictedParkingByLaw.objects.get(id=1)
+        law = ByLaw.objects.get(id=1)
         field_label = law._meta.get_field("schedule").verbose_name
         self.assertEqual(field_label, "schedule")
 
     def test_schedule_name_label(self):
-        law = RestrictedParkingByLaw.objects.get(id=1)
+        law = ByLaw.objects.get(id=1)
         field_label = law._meta.get_field("schedule_name").verbose_name
         self.assertEqual(field_label, "schedule name")
 
     def test_highway_label(self):
-        law = RestrictedParkingByLaw.objects.get(id=1)
+        law = ByLaw.objects.get(id=1)
         field_label = law._meta.get_field("highway").verbose_name
         self.assertEqual(field_label, "highway")
 
     def test_side_label(self):
-        law = RestrictedParkingByLaw.objects.get(id=1)
+        law = ByLaw.objects.get(id=1)
         field_label = law._meta.get_field("side").verbose_name
         self.assertEqual(field_label, "side")
 
     def test_between_label(self):
-        law = RestrictedParkingByLaw.objects.get(id=1)
+        law = ByLaw.objects.get(id=1)
         field_label = law._meta.get_field("between").verbose_name
         self.assertEqual(field_label, "between")
 
-    def test_times_and_or_day_label(self):
-        law = RestrictedParkingByLaw.objects.get(id=1)
+    def test_times_and_or_days_label(self):
+        law = ByLaw.objects.get(id=1)
         field_label = law._meta.get_field("times_and_or_days").verbose_name
         self.assertEqual(field_label, "times and or days")
 
-    def test_max_period_permitted_label(self):
-        law = RestrictedParkingByLaw.objects.get(id=1)
-        field_label = law._meta.get_field("max_period_permitted").verbose_name
-        self.assertEqual(field_label, "max period permitted")
-
     def test_boundary_start_label(self):
-        law = RestrictedParkingByLaw.objects.get(id=1)
+        law = ByLaw.objects.get(id=1)
         field_label = law._meta.get_field("boundary_start").verbose_name
         self.assertEqual(field_label, "boundary start")
 
     def test_boundary_end_label(self):
-        law = RestrictedParkingByLaw.objects.get(id=1)
+        law = ByLaw.objects.get(id=1)
         field_label = law._meta.get_field("boundary_end").verbose_name
         self.assertEqual(field_label, "boundary end")
 
     def test_str_method(self):
-        law = RestrictedParkingByLaw.objects.get(id=1)
+        law = ByLaw.objects.get(id=1)
         self.assertEqual(law.__str__(), "queen street (North) - 1")
+
+    def test_max_period_permitted_label(self):
+        law = ByLaw.objects.get(id=1)
+        field_label = law._meta.get_field("max_period_permitted").verbose_name
+        self.assertEqual(field_label, "max period permitted")
 
 
 class HighwayModelTest(TestCase):
