+++ parking/whereToPark/models.py
@@ -1,5 +1,4 @@
 from django.db import models
-from polymorphic.models import PolymorphicModel
 
 
 STREET_SIDES = (("W", "West"), ("E", "East"), ("N", "North"), ("S", "South"))
@@ -29,13 +28,14 @@ class Intersection(models.Model):
         unique_together = ["main_street", "cross_street"]
 
 
-class ByLaw(PolymorphicModel):
+class ByLaw(models.Model):
     """
     Polymorphic model used (abstract base class provided by django-polymorphic) in both NoParking
     and RestrictedParking models. This allows us to share fields across models
     and setting it to be abstract ensures a DB table is not created.
     """
 
+    source_id = models.IntegerField(null=True)
     schedule = models.CharField(max_length=10)
     schedule_name = models.CharField(max_length=100)
     highway = models.ForeignKey("Highway", on_delete=models.CASCADE)
@@ -50,34 +50,14 @@ class ByLaw(PolymorphicModel):
         "Intersection", on_delete=models.CASCADE, related_name="boundary_end", null=True
     )
     between = models.CharField(max_length=200, null=True)
-
-
-class NoParkingByLaw(ByLaw):
-    """
-    This model represents the "No Parking" dataset provided by the city of Toronto". Each
-    record will represent a bylaw that specifies where and when you can't park.
-    """
-
-    source_id = models.IntegerField(unique=True)
-    prohibited_times_and_or_days = models.CharField(max_length=200, null=True)
-
-    def __str__(self):
-        return f"{self.highway} ({self.side}) - {self.source_id}"
-
-
-class RestrictedParkingByLaw(ByLaw):
-    """
-    This model represents the "Restricted Parking" dataset provided by the city of Toronto. Each
-    record will represent a bylaw that specifies when/where someone is allowed to park and for
-    how long.
-    """
-
-    source_id = models.IntegerField(unique=True)
     times_and_or_days = models.CharField(max_length=200, null=True)
     max_period_permitted = models.CharField(max_length=100, null=True)
 
     def __str__(self):
         return f"{self.highway} ({self.side}) - {self.source_id}"
+
+    class Meta:
+        unique_together = ["schedule", "source_id"]
 
 
 class Highway(models.Model):
